+++ src/lib.rs
@@ -2,5 +2,6 @@ pub mod config;
 pub mod constants;
 pub mod helper;
 pub mod model;
+pub mod presenter;
 pub mod repository;
 pub mod service;
